add e2e test for error collector plugin.
add e2e test for error collector plugin.

fix comments

fix comments
