+++ core-plugins/src/e2e-test/java/io/cdap/plugin/joiner/stepsdesign/Joiner.java
@@ -15,33 +15,12 @@
  */
 package io.cdap.plugin.joiner.stepsdesign;
 
-import com.google.cloud.storage.Blob;
-import com.google.cloud.storage.StorageException;
 import io.cdap.e2e.pages.locators.CdfPluginPropertiesLocators;
 import io.cdap.e2e.utils.CdfHelper;
 import io.cdap.e2e.utils.ElementHelper;
 import io.cdap.e2e.utils.PluginPropertyUtils;
-import io.cdap.e2e.utils.StorageClient;
 import io.cdap.plugin.joiner.actions.JoinerActions;
 import io.cucumber.java.en.Then;
-import org.apache.directory.api.util.Strings;
-import org.junit.Assert;
-import stepsdesign.BeforeActions;
-
-import java.io.BufferedReader;
-import java.io.ByteArrayInputStream;
-import java.io.IOException;
-import java.io.InputStream;
-import java.io.InputStreamReader;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.nio.file.Paths;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Objects;
-
-import static io.cdap.plugin.common.stepsdesign.TestSetupHooks.gcsTargetBucketName;
-
 /**
  *  Joiner Related Step Design.
  */
@@ -56,9 +35,9 @@ public class Joiner implements CdfHelper {
     JoinerActions.uncheckPluginFieldAliasCheckBox(plugin, field);
   }
 
-  @Then("Enter numPartitions {int}")
-  public void openJoinerProperties(int numPartitions) {
-    JoinerActions.enterNumPartitions(String.valueOf(numPartitions));
+  @Then("Enter numPartitions {string}")
+  public void openJoinerProperties(String partitions) {
+    JoinerActions.enterNumPartitions(PluginPropertyUtils.pluginProp(partitions));
   }
 
   @Then("Select joiner type {string}")
@@ -69,64 +48,5 @@ public class Joiner implements CdfHelper {
   @Then("Scroll to validation button and click")
   public void scrollToValidationButton() {
     ElementHelper.clickUsingActions(CdfPluginPropertiesLocators.validateButton);
-  }
-
-  @Then("Verify the CSV Output File matches the Expected Output File With Expected Partition")
-  public void verifyCSVOutput() {
-    try {
-      // The output gcs folder will be like:
-      // e2e-test-[uuid]
-      // --2022-06-26-00-27/
-      // ----_SUCCESS
-      // ----part-r-0000
-      // ----part-r-0001
-      // ----part-r-...
-      // The number of part-r-* files should match the expected partitions.
-      int partitions = 0;
-      List<String> lst = new ArrayList<>();
-      for (Blob blob : StorageClient.listObjects(gcsTargetBucketName).iterateAll()) {
-        String name = blob.getName();
-        if (name.contains("part-r")) {
-          partitions++;
-          try (InputStream inputStream = new ByteArrayInputStream(blob.getContent())) {
-            readInputStream(inputStream, lst);
-          }
-        }
-      }
-      Path path = Paths.get(Objects.requireNonNull(Joiner.class.getResource
-        ("/" + PluginPropertyUtils.pluginProp("joinerOutput"))).getPath()).toAbsolutePath();
-      Assert.assertEquals("Output partition should match",
-                          partitions, Integer.parseInt(PluginPropertyUtils.pluginProp("expectedPartitions")));
-      Assert.assertTrue("Output content should match",
-                        Strings.equals(getSortedCSVContent(lst), new String(Files.readAllBytes(path))));
-    } catch (StorageException | IOException e) {
-      if (e.getMessage().contains("The specified bucket does not exist")) {
-        BeforeActions.scenario.write("GCS Bucket " + gcsTargetBucketName + " does not exist.");
-      } else {
-        Assert.fail(e.getMessage());
-      }
-    }
-  }
-  private String getSortedCSVContent(List<String> lst) {
-    // Since the spark output files aren't guaranteed to be ordered, the output entry with [id, first_name, last_name]
-    // schema needs to be sorted by id for comparison purpose.
-    lst.sort((s1, s2) -> {
-      String id1 = s1.split(",")[0], id2 = s2.split(",")[0];
-      return Integer.parseInt(id1) - Integer.parseInt(id2);
-    });
-    StringBuilder sb = new StringBuilder();
-    for (String s : lst) {
-      sb.append(s);
-    }
-    return sb.toString();
-  }
-
-  private void readInputStream(InputStream input, List<String> lst) throws IOException {
-    try (BufferedReader reader = new BufferedReader(new InputStreamReader(input))) {
-      String line;
-      while ((line = reader.readLine()) != null) {
-        lst.add(line + "\n");
-      }
-    }
   }
 }
