+++ core-plugins/src/e2e-test/resources/pluginParameters.properties
@@ -8,5 +8,21 @@ firstNameCsvFile=testdata/first_name.csv
 lastNameCsvFile=testdata/last_name.csv
 firstNameOutputSchema=[{"key":"id","value":"int"},{"key":"first_name","value":"string"}]
 lastNameOutputSchema=[{"key":"id","value":"int"},{"key":"last_name","value":"string"}]
+errorCollectorOutputSchema=[{"key":"id","value":"int"},{"key":"first_name","value":"string"},\
+  {"key":"msg","value":"string"},{"key":"code","value":"int"},{"key":"node","value":"string"}]
 joinerOutput=testdata/joiner_output.csv
-expectedPartitions=1
+errorCollectorOutput=testdata/error_collector_output.csv
+expectedJoinerOutputPartitions=1
+# Set the errorcollector partition to 0 since the plugin doesn't specify the output partition.
+# By setting it to 0, all partitions will be combined and sorted during the verification process.
+expectedErrorCollectorOutputPartitions=0
+javascriptTransformScript=function transform(input, emitter, context) { \
+                    if (input.id < 10) { \
+                        emitter.emitError({ \
+                            'errorCode': 1, \
+                            'errorMsg': 'blacklisted id', \
+                            'invalidRecord': input \
+                        }); \
+                        return \
+                    } \
+                }
