+++ core-plugins/src/e2e-test/resources/testdata/error_collector_output.csv
@@ -1,0 +1,9 @@
+1,Crichton,blacklisted id,1,JavaScript
+2,Monty,blacklisted id,1,JavaScript
+3,Terese,blacklisted id,1,JavaScript
+4,Mozes,blacklisted id,1,JavaScript
+5,Christian,blacklisted id,1,JavaScript
+6,Caro,blacklisted id,1,JavaScript
+7,Maye,blacklisted id,1,JavaScript
+8,Bernie,blacklisted id,1,JavaScript
+9,Heidie,blacklisted id,1,JavaScript
